Refactor: Refactor directory structure of the foods-feature-specific components (#168)
* refactor: unify component imports

* refactor: directory structure for food add/edit form

* refactor: remove unnecessary example component

* refactor: refactor directory structure for BadgeList and FilterButton

* refactor: fix relative imports

* refactor: fix indexing
