+++ src/features/foods/components/Form/index.ts
@@ -1,6 +1,3 @@
-export { AddButton } from './AddButton';
-export { AddDrawerBody } from './AddDrawerBody';
 export { AddDrawerContent } from './AddDrawerContent';
 export { CategorySelect } from './CategorySelect';
-export { EditDrawerBody } from './EditDrawerBody';
 export { EditDrawerContent } from './EditDrawerContent';
